% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_Kb.R
\name{calculate_baseline_effective_conductivity}
\alias{calculate_baseline_effective_conductivity}
\alias{calculate_Kb}
\title{Baseline effective conductivity estimation  (Kb)}
\source{
page 14 of \url{https://www.ars.usda.gov/ARSUserFiles/50201000/WEPP/usersum.pdf}
}
\usage{
calculate_baseline_effective_conductivity(sand, clay, cec)

calculate_Kb(sand, clay, cec)
}
\arguments{
\item{sand}{percent of sand}

\item{clay}{percent of clay}

\item{cec}{cation exchange capacity of the soil (meq/100g)}
}
\value{
numeric, Kb, the baseline effective conductivity
}
\description{
This function will calculation the baseline effective conductivity based on
the amount of sand, clay and the cation exchange capacity of the soil.
}
\section{Functions}{
\itemize{
\item \code{calculate_Kb}: 
}}

\examples{
calculate_baseline_effective_conductivity(5.2, 40.1, 29.4)
}
\seealso{
\link{effective_hydraulic_conductivity}
}
