% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soil.R
\name{calculate_cum_dist}
\alias{calculate_cum_dist}
\title{Calculate cumulative distance across groups used for linearizing the soil file}
\usage{
calculate_cum_dist(layer, distance)
}
\arguments{
\item{layer}{the layer column in sol object}

\item{distance}{the distance column in sol object}
}
\value{
a numeric vector containing cumulative distance
}
\description{
Calculate cumulative distance across groups used for linearizing the soil file
}
\examples{
calculate_cum_dist(sol$layer, sol$distance)

}
\seealso{
\code{\link{plot.Soil}}
}
