% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slope.R
\name{calculate_diff_slope}
\alias{calculate_diff_slope}
\title{Calculates difference in OFE rows}
\usage{
calculate_diff_slope(slp_df)
}
\arguments{
\item{slp_df}{A slope data frame}
}
\value{
Sum of the lagged difference of slopes and OFEs
}
\description{
OFEs have matching slope as it transitions from one OFE to another. To validate
this property, we calculate the difference of the last slope of an OFE and the first
slope of the consecutive OFE and check whether this difference is zero.
}
\examples{
calculate_diff_slope(slp_df)
}
