% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_Ke.R
\name{calculate_Ke}
\alias{calculate_Ke}
\alias{calculate_Kef}
\title{Calculate time-invariant effective hydraulic conductivity (Ke-constant)}
\source{
page 26 of \url{https://www.ars.usda.gov/ARSUserFiles/50201000/WEPP/usersum.pdf}
}
\usage{
calculate_Ke(hydrologic_group, management, sand = NULL)

calculate_Kef(hydrologic_group, sand)
}
\arguments{
\item{hydrologic_group}{hydrologic soil group (A,B,C,D)}

\item{management}{management practice}

\item{sand}{}
}
\description{
Calculate time-invariant effective hydraulic conductivity (Ke-constant)
}
\section{Functions}{
\itemize{
\item \code{calculate_Kef}: fallow soil Ke
}}

