% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Slope.R
\name{calculate_total_distance}
\alias{calculate_total_distance}
\title{Calculate the total distance}
\usage{
calculate_total_distance(slp)
}
\arguments{
\item{slp}{A slope object}
}
\value{
A numeric vector of total distances
}
\description{
By default, Slope objects include only the distance for each OFE rather
than the total distance in the hillslope. This function calculates the
total distance.
}
\examples{
slp <- read_slp(system.file("extdata", "071000090603_2.slp", package="WEPPR"))
calculate_total_distance(slp)

}
\seealso{
\code{\link{calculate_total_distance}}, \code{\link{integrate_slp}}, \code{\link{expand_slp}}, \code{\link{remove_slp_transitions}}
}
