% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Climate.R
\name{Climate}
\alias{Climate}
\alias{new_Climate}
\title{Create a Climate Class}
\usage{
new_Climate(cli = list())
}
\arguments{
\item{cli}{A list of data frames}
}
\value{
Climate object with class \code{Climate} and \code{data.frame}
}
\description{
The climate class constructor appends the Climate class to the climate data frame.
The constructor validates the climate data frame, and the climate class provides
functionality to plot the instantaneous and cumulative precipitation,
and to interpolate the climate data
}
\examples{
fpath_cli <- system.file("extdata", "092.63x040.90.cli", package="WEPPR")
(cli <- read_cli(fpath_cli))
}
