% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_run.R
\name{create_run}
\alias{create_run}
\title{Create a run object}
\usage{
create_run()
}
\value{
A data frame containing two columns
\describe{
  \item{name}{names of variables}
  \item{value}{the value for that variable}
}
}
\description{
Interactively create a run object by asking questions in exactly the same
order as the WEPP interactive system. Ideally, you only need to go through
this process once and then you can modify the run object as needed, e.g. to
change file names.
}
\details{
This run object specifies the type of WEPP simulation to perform as well as
names for input/output files.
}
