% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_Kb.R
\docType{data}
\name{effective_hydraulic_conductivity}
\alias{effective_hydraulic_conductivity}
\title{Effective hydraulic conductivity}
\format{
A data frame with 43 rows and 10 variables
\describe{
  \item{soil}{character, soil type}
  \item{sand}{numeric, sand content (\%)}
  \item{clay}{numeric, clay content (\%)}
  \item{organic matter}{numeric, organic matter content (\%)}
  \item{cec}{numeric, cation exchange capacity (meq/100g)}
  \item{simulator measured Ke}{numeric, simulator measured Ke (mm/hr)}
  \item{opt. constant Kef}{numeric, optimized constant Kef (mm/hr)}
  \item{est. constant Kef}{numeric, estimated constant Kef (mm/hr)}
  \item{opt. baseline Kb}{numeric, optimized baseline Kb (mm/hr)}
  \item{est. baseline Kb}{numeric, estimated baseline Kb (mm/hr)}
}
}
\usage{
effective_hydraulic_conductivity
}
\description{
Effective hydraulic conductivity
}
\seealso{
\link{calculate_baseline_effective_conductivity}
}
\keyword{datasets}
