% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Climate.R
\name{expand_cli}
\alias{expand_cli}
\title{Expands Climate object to calculate ppt and inst ppt. at specific hours}
\usage{
expand_cli(cli, n = 1001)
}
\arguments{
\item{cli}{A Slope object}

\item{n}{(optional) a positive integer indicating how many equally-spaced time to
calculate.}
}
\value{
a data.frame containing time, ppt, and inst. ppt
}
\description{
Expands Climate object to calculate ppt and inst ppt. at specific hours
}
\examples{
cli <- read_cli(system.file("extdata", "092.63x040.90.cli", package="WEPPR"))\
expand_cli(cli)

}
\seealso{
\code{\link{plot.Slope}}, \code{\link{integrate_cli}}
}
