% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Slope.R
\name{expand_slp}
\alias{expand_slp}
\title{Expand Slope object to calculate slope and elevation at specific distances}
\usage{
expand_slp(slp, distances = NULL, n = 1001)
}
\arguments{
\item{slp}{A Slope object}

\item{distances}{a numeric vector of distances}

\item{n}{a positive integer indicating how many equally-spaced distances to
calculate. Ignored if \code{distances} is provided.}
}
\value{
a data.frame containing distance, slope, and elevation
}
\description{
This function will calculate the slope and elevation of a Slope object for
specific distances
}
\examples{
slp <- read_slp(system.file("extdata", "071000090603_2.slp", package="WEPPR"))
expand_slp(slp)

}
\seealso{
\code{\link{plot_Slope}}, \code{\link{integrate_slp}}, \code{\link{remove_slp_transitions}}
}
