% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soil.R
\name{get_dup_count}
\alias{get_dup_count}
\title{Calculate the duplicate count for each row based on solthk bin for linearizing
soil file}
\usage{
get_dup_count(sol)
}
\arguments{
\item{sol}{A soil object}
}
\value{
A numeric vector of duplicate counts
}
\description{
Calculate the duplicate count for each row based on solthk bin for linearizing
soil file
}
\examples{
dup_counts <- get_dup_count(sol)
dup_counts

}
\seealso{
\code{\link{linearize_sol}}
}
