% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soil.R
\name{get_xmin}
\alias{get_xmin}
\title{Calculate xmin vector for geom_rect}
\usage{
get_xmin(layer, distance)
}
\arguments{
\item{layer}{the layer column in sol object}

\item{distance}{the distance column in sol object}
}
\value{
a numeric vector xmin needed to plot geom_rect
}
\description{
Calculate xmin vector for geom_rect
}
\examples{
get_xmin(sol$layer, sol$distance)

}
\seealso{
\code{\link{plot.Soil}}
}
