% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.R
\name{hash}
\alias{hash}
\title{Compute md5sum hash for a file}
\usage{
hash(files)
}
\arguments{
\item{files}{path and names of files to calculate md5sum}
}
\value{
md5sum of file
}
\description{
Currently just wraps tools::md5sum for convenience.
}
\details{
Perhaps in the future it will also hash R objects perhaps by writing them to
a file and then hashing the file.
}
