% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Climate.R
\name{integrate_cli}
\alias{integrate_cli}
\title{Calculates the instantaneous precipitation (precipitation / time in hours)}
\usage{
integrate_cli(cli)
}
\arguments{
\item{cli}{A Climate object}
}
\value{
A \code{data.frame} that is a modified version of the Climate object
with additional columm m
}
\description{
Calculates the instantaneous precipitation (precipitation / time in hours)
}
\examples{
cli <- read_cli(system.file("extdata", "092.63x040.90.cli", package="WEPPR"))
integrate_cli(cli)

}
\seealso{
\code{\link{plot.Climate}}, \code{\link{expand_cli}}
}
