% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_wepp_available.R
\name{is_wepp_available}
\alias{is_wepp_available}
\title{Will check for WEPP executable}
\usage{
is_wepp_available()
}
\value{
if WEPP executable is in the path, returns TRUE otherwise returns FALSE
}
\description{
Will check for WEPP executable
}
