% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slope.R
\name{linearize_slp}
\alias{linearize_slp}
\title{Linearize the slope data file}
\usage{
linearize_slp(slp_class)
}
\arguments{
\item{slp_class}{A slope object}
}
\description{
Linearize the slope data file
}
\examples{
slp <- read_slp(system.file("extdata", "071000090603_2.slp", package="WEPPR"))
linearized_slp <- linearize_slp(slp)
}
