% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soil.R
\name{linearize_sol}
\alias{linearize_sol}
\title{Linearize the soil data file}
\usage{
linearize_sol(slp_sol)
}
\arguments{
\item{slp_sol}{A Soil object combined with slope data}
}
\value{
a one-row dataframe containing soil data
}
\description{
Converts the soil file into one WEPP run (data frame with one row)
}
\examples{
slp_sol <- merge_slp_sol(slp, sol)
lin_sol <- linearize_sol(slp_sol)

}
\seealso{
\code{\link{expand_slp}}
}
