% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soil.R
\name{merge_slp_sol}
\alias{merge_slp_sol}
\title{Merge Slope and Soil object to provide Soil object with distance for plotting}
\usage{
merge_slp_sol(slp, sol)
}
\arguments{
\item{layer}{the layer column in sol object}

\item{distance}{the distance column in sol object}
}
\value{
slp_sol object with class Soil and tibble
}
\description{
Merge Slope and Soil object to provide Soil object with distance for plotting
}
\examples{
fpath_slp <- system.file("extdata", "071000090603_2.slp", package="WEPPR")
fpath_sol <- system.file("extdata", "071000090603_2.sol", package="WEPPR")
slp <- read_slp(fpath_slp)
sol <- read_sol(fpath_sol)
slp_sol <- merge_slp_sol(slp, sol)
slp_sol

}
\seealso{
\code{\link{plot.Soil}}
}
