% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slope.R
\name{new_Slope}
\alias{new_Slope}
\title{Create a Slope Class}
\usage{
new_Slope(x = data.frame())
}
\arguments{
\item{file}{A path to the slp file}
}
\description{
Creates a slope class from a slope (*.slp) file
}
\examples{
example
slp_class <- new_Slope(slp_data)
validate_Slope(slp_class)
plot(slp_class)
linearize_slp(slp_class)
}
