% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Climate.R
\name{plot.Climate}
\alias{plot.Climate}
\title{Plots the ppt and inst. ppt from a Climate object}
\usage{
\method{plot}{Climate}(cli, n = 1001, Date = NULL, plots = c("ppt_inst", "ppt_cum"))
}
\arguments{
\item{cli}{a Climate object}

\item{n}{(optional) a positive integer indicating how many points to evaluate for drawing}

\item{plots}{(optional) a charcter vector indicating whether to draw the "ppt_inst", "ppt_cum", or both}

\item{date}{a date string value with format "YYYY-MM-DD"}
}
\value{
a gtable containing plots of slope or elevation
}
\description{
Plots the ppt and inst. ppt from a Climate object
}
\examples{
cli <- read_cli(system.file("extdata", "092.63x040.90.cli", package="WEPPR"))
plot(cli)
plot(cli, date = "2017-09-21", plots=c("ppt_inst"))
plot(cli, date = "2017-09-21", plots=c("ppt_cum"))
plot(cli, date = "2017-09-21")

}
\seealso{
\code{\link{expand_cli}}
}
