% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slope.R
\name{plot.Slope}
\alias{plot.Slope}
\title{Plots the slope and elevation from a Slope object}
\usage{
\method{plot}{Slope}(slp, n = 1001, plots = c("slope", "elevation"))
}
\arguments{
\item{slp}{a Slope object}

\item{n}{(optional) a positive integer indicating how many points to evaluate for drawing}

\item{plots}{(optional) a charcter vector indicating whether to draw the "slope", "elevation", or both}
}
\value{
a gtable containing plots of slope or elevation
}
\description{
The Slope object contains distances and slopes. The remainder of the slopes
need to be linearly interpolated. These linearly interpreted slopes are the
derivative of the elevation (hillslope) profile. We start each hillslope at
the point (0,0).
}
\examples{
slp <- read_slp(system.file("extdata", "071000090603_2.slp", package="WEPPR"))
plot(slp, plots=c("slope"))
plot(slp, plots=c("elevation"))
plot(slp)

}
