% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soil.R
\name{plot.Soil}
\alias{plot.Soil}
\title{Plots the features in the Soil object}
\usage{
\method{plot}{Soil}(slp_sol)
}
\arguments{
\item{slp_sol}{a Soil object merged with Slope data}
}
\value{
multiple ggplot objects which are plots of soil features
}
\description{
Plots the features in the Soil object
}
\examples{
slp_sol <- merge_slp_sol(slp, sol)
plot(slp_sol)

}
