% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cli.R
\name{read_cli}
\alias{read_cli}
\title{Read a WEPP climate file}
\usage{
read_cli(file)
}
\arguments{
\item{file}{A path to the file.}
}
\description{
Reads a Water Erosion Prediction Project (WEPP) climate (*.cli) file.
This file contains climate information.
}
