% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dep_run.R
\name{read_dep_run}
\alias{read_dep_run}
\title{Read a WEPP run file in DEP format}
\usage{
read_dep_run(file)
}
\arguments{
\item{file}{A path to the file.}
}
\value{
A character vector with the following named elements:
\describe{
  \item{man}{management input filename}
  \item{slp}{slope input filename}
  \item{sol}{soil input filename}
  \item{cli}{climate input filename}
  \item{wb}{wb output filename}
  \item{env}{env output filename}
  \item{yld}{yield output filename}
}
}
\description{
This is a temporary function until we determine how to deal with run files
globally. This function will read in a run file in the daily erosion project
(DEP) format. It will only record the input and ouput files.
}
