% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_env.R
\name{read_env}
\alias{read_env}
\title{Read a WEPP environment file}
\usage{
read_env(file)
}
\arguments{
\item{file}{A path to the file.}
}
\description{
Reads a Water Erosion Prediction Project (WEPP) environment (*.env) file.
This file contains output from a WEPP run
}
