% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_run.R
\name{read_run}
\alias{read_run}
\title{Read a WEPP run file}
\usage{
read_run(file)
}
\arguments{
\item{file}{A path to the file.}
}
\description{
Reads a Water Erosion Prediction Project (WEPP) run (*.run) file.
This file contains run settings (?) information.
}
