% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_slp.R
\name{read_sol}
\alias{read_sol}
\title{Read a WEPP soil file}
\usage{
read_sol(file)
}
\arguments{
\item{file}{A path to the file.}
}
\description{
Reads a Water Erosion Prediction Project (WEPP) soil (*.sol) file.
This file contains soil information.
}
