% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wb.R
\name{read_wb}
\alias{read_wb}
\title{Read a WEPP wb file}
\usage{
read_wb(file)
}
\arguments{
\item{file}{A path to the file.}
}
\value{
A data.frame with the columns:
\describe{
  \item{ofe}{numeric, overland flow element}
  \item{jday}{numeric (1-366), julian day}
  \item{year}{numeric, YYYY}
  \item{precip}{numeric, amount of precipitation}
  \item{runoff}{numeric, amount of runoff}
  \item{sw}{numeric}
  \item{sw1}{numeric}
  \item{sw2}{numeric}
  \item{ep}{numeric}
  \item{es}{numeric}
  \item{er}{numeric}
}
}
\description{
Reads a Water Erosion Prediction Project (WEPP) soil water balance (*.wb) file.
This file contains soil water balance information.
}
