% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_run.R
\docType{data}
\name{run_variables}
\alias{run_variables}
\title{Run option variables}
\format{
An object of class \code{data.frame} with 28 rows and 2 columns.
}
\usage{
run_variables
}
\description{
Contains all possible run option variables with the associated question
that is asked.
}
\keyword{datasets}
