% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soil.R
\name{Soil}
\alias{Soil}
\alias{new_Soil}
\title{Create a Soil Class}
\usage{
new_Soil(sol = data.frame())
}
\arguments{
\item{sol}{A soil dataframe}
}
\value{
Soil object with class \code{Soil} and \code{data.frame}
}
\description{
The soil class constructor appends the Soil class to the Soil object
The constructor validates the soil data frame, and the Soil class provides
functionality to plot the soil features and to linearize the Soil data
}
\examples{
fpath_sol <- system.file("extdata", "071000090603_2.sol", package="WEPPR")
sol <- read_sol(fpath_sol)
}
