% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cli.R
\name{write_cli}
\alias{write_cli}
\title{Write a WEPP climate file}
\usage{
write_cli(cli, path)
}
\arguments{
\item{cli}{A \code{cli} object.}

\item{path}{Path to write to.}
}
\description{
Writes a Water Erosion Prediction Project (WEPP) climate (*.cli) file.
This file contains climate information.
}
