% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dep_run.R
\name{write_dep_run}
\alias{write_dep_run}
\title{Writes a run file in the DEP format}
\usage{
write_dep_run(files, run_filename)
}
\arguments{
\item{files}{A named character vector with names man, slp, sol, cli, wb, env,
and yld}

\item{run_filename}{A character for the name of the run file.}
}
\description{
This is a temporary function until we determine how to deal with run files
globally. This will write a run file using all the same options used for the
daily erosion project but replacing the input/output filenames with those
passed to the function.
}
