% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_env.R
\name{write_env}
\alias{write_env}
\title{Write a WEPP environment file}
\usage{
write_env(env, path)
}
\arguments{
\item{env}{A \code{env} object.}

\item{path}{Path to write to.}
}
\description{
Writes a Water Erosion Prediction Project (WEPP) environment (*.env) file.
}
