% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_man.R
\name{write_man}
\alias{write_man}
\title{Write a WEPP management file}
\usage{
write_man(man, path)
}
\arguments{
\item{man}{A \code{man} object.}

\item{path}{Path to write to.}
}
\description{
Writes a Water Erosion Prediction Project (WEPP) management (*.man) file.
This file contains land management information.
}
