% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_run.R
\name{write_run}
\alias{write_run}
\title{Write a WEPP run settings file}
\usage{
write_run(run, path)
}
\arguments{
\item{run}{A \code{run} object.}

\item{path}{Path to write to.}
}
\description{
Writes a Water Erosion Prediction Project (WEPP) run settings (*.run) file.
This file contains run settings.
}
