% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_slp.R
\name{write_slp}
\alias{write_slp}
\title{Write a WEPP slope file}
\usage{
write_slp(slp, path)
}
\arguments{
\item{slp}{A \code{slp} object.}

\item{path}{Path to write to.}
}
\description{
Writes a Water Erosion Prediction Project (WEPP) slope (*.slp) file.
}
