% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slope.R
\name{write_slp_file}
\alias{write_slp_file}
\title{Write the slope data file}
\usage{
write_slp_file(slp, path)
}
\arguments{
\item{slp}{A \code{slp} object.}

\item{path}{Path to write to.}
}
\description{
Write the slope data file
}
\examples{
slp <- read_slp(system.file("extdata", "071000090603_2.slp", package="WEPPR"))
write_slp(slp, "example.slp")

}
