% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sol.R
\name{write_sol}
\alias{write_sol}
\title{Write a WEPP soil file}
\usage{
write_sol(sol, path)
}
\arguments{
\item{sol}{A \code{sol} object.}

\item{path}{Path to write to.}
}
\description{
Writes a Water Erosion Prediction Project (WEPP) soil (*.sol) file.
This file contains soil information.
}
