% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soil.R
\name{write_sol_file}
\alias{write_sol_file}
\title{Write the soil data file}
\usage{
write_sol_file(sol, path)
}
\arguments{
\item{sol}{A \code{sol} object.}

\item{path}{Path to write to.}
}
\description{
Write the soil data file
}
\examples{
#' fpath_sol <- system.file("extdata", "071000090603_2.sol", package="WEPPR")
sol <- read_sol(fpath_sol)
write_sol(sol, "example.sol")

}
