% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_wb.R
\name{write_wb}
\alias{write_wb}
\title{Write a WEPP wb file}
\usage{
write_wb(wb, path)
}
\arguments{
\item{wb}{A \code{wb} object.}

\item{path}{Path to write to.}
}
\description{
Writes a Water Erosion Prediction Project (WEPP) soil water balance (*.wb) file.
This file contains soil water balance information.
}
