% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_yld.R
\name{write_yld}
\alias{write_yld}
\title{Write a WEPP yield file}
\usage{
write_yld(yld, path)
}
\arguments{
\item{yld}{A \code{yld} object.}

\item{path}{Path to write to.}
}
\description{
Writes a Water Erosion Prediction Project (WEPP) yield (*.yld) file.
This file contains yield information.
}
