% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slope.R
\name{linearize_slp}
\alias{linearize_slp}
\title{Linearize the slope data file}
\usage{
linearize_slp(slp, n = 1001)
}
\arguments{
\item{slp}{A \code{slp} object}

\item{n}{(optional) a positive integer indicating how many slope columns to evaluate}
}
\value{
a one-row dataframe containing total distance and slope columns
}
\description{
Converts the slope file into one WEPP run (data frame with one row)
}
\examples{
slp <- read_slp(system.file("extdata", "071000090603_2.slp", package="WEPPR"))
lin_slp <- linearize_slp(slp)

}
\seealso{
\code{\link{expand_slp}}
}
