% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slope.R
\name{new_Slope}
\alias{new_Slope}
\title{Create a Slope Class}
\usage{
new_Slope(x = data.frame())
}
\arguments{
\item{dataframe}{A slope dataframe}
}
\description{
Creates a Slope class from a slope data frame
Validates the slope file for matching OFEs
Plots the slope plot and elevation plot
Linearize the slope file into one WEPP run
}
\examples{
slp_class <- new_Slope(slp_data)
validate_Slope(slp_class)
plot(slp_class)
linearize_slp(slp_class)
}
