% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soil.R
\name{plot.Soil}
\alias{plot.Soil}
\title{Plots the features in the Soil object}
\usage{
\method{plot}{Soil}(slp_sol, types = c("all"))
}
\arguments{
\item{slp_sol}{a Soil object merged with Slope data}

\item{type}{a list of types to include in the plot}
}
\value{
multiple ggplot objects which are plots of soil features
}
\description{
Plots the features in the Soil object
}
\examples{
slp_sol <- merge_slp_sol(slp, sol)
plot(slp_sol, type=c("all"))
plot(slp_sol, type=c("salb", "kt", "kr"))

}
