% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_env.R
\name{read_env}
\alias{read_env}
\title{Read a WEPP environment file}
\usage{
read_env(file, start.year)
}
\arguments{
\item{file}{A path to the file.}

\item{start.year}{First year of simulation.}
}
\value{
A \code{env} \code{data.frame} with the following columns:
\describe{
  \item{day}{Numeric, day of the month}
  \item{mo}{Numeric, day of the month}
  \item{year}{Numeric, day of the month}
  \item{Precp}{Numeric, day of the month}
  \item{Runoff}{Numeric, day of the month}
  \item{IR-det}{Numeric, day of the month}
  \item{Av-det}{Numeric, day of the month}
  \item{Mx-det}{Numeric, day of the month}
  \item{Point}{Numeric, day of the month}
  \item{Av-dep}{Numeric, day of the month}
  \item{Max-dep}{Numeric, day of the month}
  \item{Point}{Numeric, day of the month}
  \item{Sed.Del}{Numeric, sediment delivery ratio?}
  \item{ER}{Numeric, enrichment ratio?}
}
}
\description{
Reads a Water Erosion Prediction Project (WEPP) environment (*.env) file.
This file contains environmental output from
}
