% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_man.R
\name{read_man}
\alias{read_man}
\title{Read a WEPP management file}
\usage{
read_man(file)
}
\arguments{
\item{file}{A path to the file.}
}
\description{
Reads a Water Erosion Prediction Project (WEPP) plant/management (*.man) file.
This file contains plant/management information.
}
