% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sol.R
\name{read_sol}
\alias{read_sol}
\title{Read a WEPP soil file}
\usage{
read_sol(file)
}
\arguments{
\item{file}{A path to the file.}
}
\value{
A \code{Soil} object with two classes - \code{Soil} and \code{data.frame}
}
\description{
Reads a Water Erosion Prediction Project (WEPP) soil (*.sol) file.
This file contains soil information.
}
\examples{
fpath_sol <- system.file("extdata", "071000090603_2.sol", package="WEPPR")
sol <- read_sol(fpath_sol)

}
