% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wb.R
\name{read_wb}
\alias{read_wb}
\title{Read a WEPP wb file}
\usage{
read_wb(file)
}
\arguments{
\item{file}{A path to the file.}
}
\description{
Reads a Water Erosion Prediction Project (WEPP) wb (*.wb) file.
This file contains water balance (?) information.
}
