% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_yld.R
\name{read_yld}
\alias{read_yld}
\title{Read a WEPP yield file}
\usage{
read_yld(file)
}
\arguments{
\item{file}{A path to the file.}
}
\description{
Reads a Water Erosion Prediction Project (WEPP) yield (*.yld) file.
This file contains yield information.
}
