% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_wepp.R
\name{run_wepp}
\alias{run_wepp}
\title{Runs the WEPP model}
\usage{
run_wepp(file)
}
\arguments{
\item{file}{A path to the file.}
}
\value{
A data frame containing the following elements:
\describe{
  \item{file}{character, file name}
  \item{file_type}{character, type of file eg. climate, environment,..}
  \item{type}{character, input/output}
  \item{md5sum}{character, hash of the file}
  \item{id}{character, conconated hash of hash of all the file names}
}
}
\description{
An R interface to run the Water Erosion Prediction Project (WEPP) model.
}
