% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slope.R
\name{Slope}
\alias{Slope}
\alias{new_Slope}
\title{Create a Slope Class}
\usage{
new_Slope(slp = data.frame())
}
\arguments{
\item{dataframe}{A slope dataframe}
}
\value{
Slope object with class \code{Slope} and \code{data.frame}
}
\description{
The slope class constructor appends the Slope class to the slope data frame.
The constructor validates the slope data frame, and the slope class provides
functionality to plot the elevation and slope plot, and to linearize the
slope data
}
\examples{
fpath_slp <- system.file("extdata", "071000090603_2.slp", package="WEPPR")
slp <- read_slp(fpath_slp)
}
